requirements.txt
pyelftools
pefile
capstone
